refactor: move scroll cursor actions to vim crate
Update the approach which is being used to implement both `z l` and `z
h` in vim mode so that the actions are actually defined in the vim
crate, namely `ColumnRight` and `ColumnLeft`.

It was noted that `ctrl-e` and `ctrl-y` already exist, and it is very
similar to what is trying to be achieved, with the difference that those
motions work at a line level, while `z l` and `z h` work at a column
level. However, since these are already present, the focus will now
shift towards trying to leverage the existing code and get this working
with the existin `ScrollAmount` and `ScrollDirection` structs.