feat(diagnostics): add buffer diagnostics editor
Introduce a new diagnostics view, `BufferDiagnosticsEditor`, which
allows users to view the diagnostics for the currently open file. A new
action has also been introduced so users can deploy this view from the
command palette, `diagnostics: deploy current file`.

This view is very similar in behaviour to `ProjectDiagnosticsEditor`,
but it was worth noting that, while `ProjectDiagnosticsEditor` reads and
writes to the global `IncludeWarnings` setting,
`BufferDiagnosticsEditor` only reads its value when first being created,
otherwise it keeps its own `include_warnings` state so as to not affect
the `ProjectDiagnosticsEditor` or other `BufferDiagnosticsEditor`
entities.

A new method was introduced to make it easier to retrieve the diagnostic
summary for a specific path rather for all paths,
`project::lsp_store::LspStore.diagnostic_summary_for_path`, seeing as
`BufferDiagnosticsEditor` only needs to display the summary for a single
path.