chore: save project path on deploy and re-focus tab
Update the `BufferDiagnosticsEditor` implementation with two new
changes:

- When running the `DeployCurrentFile` action, it now finds the current
project path and saves it so we can later access it and display it in
the tab's text. - When deploying a new `BufferDiagnosticsEditor`, it
first checks if there's already one for the same project path and, in
that case, updates the workspace to focus on that existing tab instead.