chore: remove path_matcher from project diagnostics editor
Remove `path_matcher` from `ProjectDiagnosticsEditor`, as well as any
other fields and methods/functions that might have been associated with
this specific functionality, as this is being replaced by the
`BufferDiagnosticsEditor` view.

The `ToolbarControls` module for the `ProjectDiagnosticsEditor` has also
been updated, so as to remove the filter filter toggle button that
interacted with the `path_matcher` field.