feat: diagnostic blocks in buffer diagnostics editor
Update the `BufferDiagnosticsEditor` view in order to mimic the existing
behavior of `ProjectDiagnosticsEditor` when it comes to displaying
blocks in the editor for each of the diagnostics.

The current implementation seems to not be correctly updating when
there's more errors being added, so it's something that still needs to
be fixed in a future commit.