fix: correctly allow file changes to be saved on buffer diagnosticso
Update the `Item` trait implementation for `BufferDiagnosticsEditor` in
order to allow the buffer's contents to be saved directly from the
editor that is rendered in the `BufferDiagnosticsEditor`. This is
achieved by implementing both the `can_save` and `save` methods from the
`Item` trait.

The `breadcrumb_location` and `breadcrumbs` methods, for the same trait,
have also been added in order to display breadcrumbs when the
`BufferDiagnosticsEditor` is being displayed.