refactor: replace usage of global include warnings setting
While reviewing the new `BufferDiagnosticsEditor`, it was noted that, if
the `ProjectDiagnosticsEditor` was deployed, and the tooltip to toggle
the warnings was pressed, it would update both the project diagnostics
editor as well as the buffer diagnostics editor. This happens because
both are reading from the gobal `IncludeWarnings` setting.

It's better for the buffer diagnostics editor to be a little bit more
independent, so this commit updates its implementation to only leverage
the global `IncludeWarnings` setting when first setting the
`include_warnings` value. Afterwards, the `toggle_warnings` method will
only update the `include_warnings` field and avoid reading or updating
the global `IncludeWarnings` setting, so no updates to the project
diagnostics editor warnings setting will have an effect on the buffer
diagnostics editor.