refactor: improve diagnostic summary for path
- Update the `project::lsp_store::LspStore.diagnostic_summary_for_path`
method to no longer rely on the `diagnostic_summaries` method and,
instead of iterating over all worktrees and paths, until the first
diagnostic summary for the desired path is found, it fetches the
available diagnostic summaries for the worktree of the provided
`ProjectPath`, and the first `DiagnosticSummary`, for any language
server. This ends up being more performant as, in the worst case, the
previous implementation could end up iterating over all worktress and
paths.
- Remove the `include_ignored` parameter from the
`project::lsp_store::LspStore.diagnostic_summary_for_path` method, as
that was only required due to the `diagnostic_summaries` method.
- Update `BufferDiagnosticsEditor` to no longer update the diagnostic
summary in a task, seeing as the new implementation should be much
faster and should not require this work to be carried out in the
background, so the `diagnostic_summary_task` field has been removed and
the handler for `Event::DiagnosticsUpdated` has been updated to simply
update the diagnostic summary straight away.
- Remove leftover comment related to previous `PathMatcher`
implementation.