refactor: open buffer only once in buffer diagnostics editor
Refactor the implementation of
`BufferDiagnosticsEditor.update_all_excerpts` in order to prevent
calling `Project.open_buffer` everytime this function is run.

Instead, a `buffer` field has been added to the
`BufferDiagnosticsEditor` struct, and a background task is used to open
the buffer in the background after the `BufferDiagnosticsEditor` is
initialized, ensuring that the buffer is only opened once.