chore: implement additional workspace item trait methods
When implementing the `workspace::item::Item` trait for the
`diagnostics::buffer_diagnostics::BufferDiagnosticsEditor` view, only
some of the methods were implemented, the minimum to ensure the
correct functionality for this new view.

However, not implementing some of the trait's methods meant that some
functionalities supported by `ProjectDiagnosticsEditor` are not
correctly supported by `BufferDiagnosticsEditor`. For example, while on
a specific excerpt, using `g space` should open the file/buffer in that
specific location in the file editor. This was not working because the
`added_to_workspace` method was not implemented.

This commit finishes implementing the same `workspace::item::Item` trait
methods in `BufferDiagnosticsEditor` that are also implemented for
`ProjectDiagnosticsEditor`.