feat(vim): introduce single whitespace handling in surround
This commit introduces a new value to the `SurroundWhiteSpace` enum,
`Single`, which is used when updating from an opening bracket to another
opening bracket, like so:

`{     a     }` -> cs{[ -> `[ a ]`

To achieve the same behavior as vim's surround implementation, this
commit updates both the `vim::CurlyBrackets` action and the
`Object::CurlyBrackets` enum value to support specifying which
character, if any, is associated with the action and the object. This
allows access to this character when determining how to handle
whitespace, as both characters need to be taken into account.

Since I'm not entirely sure this is the approach that should be taken,
as there could be another way to get the full command string and extract
the character from there, I've only implemented this change for
`CurlyBrackets` and will hold off on implementing it for other actions
and objects until there's confirmation that this is an acceptable
approach.